requirements.txt
django==2.2.5
djangorestframework==3.13.1
redis==3.5.3
PyMySQL==1.0.2
django-reversion==5.0.2